inventoryForm ready, event form on progress
without css
