+++ package.json
@@ -11,15 +11,12 @@
     "react": "^17.0.2",
     "react-datepicker": "^3.8.0",
     "react-dom": "^17.0.2",
-
     "react-icons": "^4.2.0",
     "react-redux": "^7.2.4",
     "react-router-dom": "^5.2.0",
-
     "react-scripts": "4.0.3",
     "react-select": "^4.3.0",
     "reactstrap": "^8.9.0",
-
     "redux": "^4.1.0",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.3.0",
