+++ src/redux/store.js
@@ -1,10 +1,10 @@
 import { createStore, compose, applyMiddleware } from "redux";
 import logger from "redux-logger";
 import thunk from "redux-thunk";
-import templateReducer from "./reducers/templateReducer";
+import formReducer from "./reducers/formReducer";
 
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
 export const store = createStore(
-  templateReducer,
+  formReducer,
   composeEnhancers(applyMiddleware(thunk, logger))
 );
