+++ src/redux/types/template.js
@@ -1,3 +1,4 @@
 // template
 export const POST_IMAGE_ACTION = "POST_IMAGE_ACTION";
 export const HANDLE_SUBMISSION_ACTION = "HANDLE_SUBMISSION_ACTION";
+export const HANDLE_EVENT_SUBMISSION_ACTION = "HANDLE_EVENT_SUBMISSION_ACTION";
